Collection authentication and authorization (#94)
* implement user authentication and authorization

* fix delete

* remove return value from remove function

* change exception message

+++ packages/backend/src/collection/collection.service.ts
@@ -1,41 +1,46 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common'
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common'
 import { PrismaService } from 'nestjs-prisma'
 import { CreateCollectionDto } from './dto/CreateCollection.dto'
 import { UpdateCollectionDto } from './dto/UpdateCollection.dto'
+import { User, UserRole } from '@prisma/client'
 
 @Injectable()
 export class CollectionService {
   constructor(private readonly prisma: PrismaService) {}
-  create(createCollectionDto: CreateCollectionDto) {
-    return this.prisma.collection.create({ data: createCollectionDto })
+  create(createCollectionDto: CreateCollectionDto, user: User) {
+    return this.prisma.collection.create({ data: { ...createCollectionDto, userId: user.id } })
   }
   findAll() {
     return this.prisma.collection.findMany()
   }
   async findOne(id: string) {
-    const collection = await this.prisma.collection.findUnique({
-      where: { id },
-    })
+    const collection = await this.prisma.collection.findUnique({ where: { id } })
     if (!collection) {
       throw new NotFoundException('Collection not found')
     }
     return collection
   }
-  async update(id: string, updateCollectionDto: UpdateCollectionDto) {
-    const collection = await this.prisma.collection.update({
+  async update(id: string, updateCollectionDto: UpdateCollectionDto, user: User) {
+    const collection = await this.prisma.collection.findUnique({ where: { id } })
+    if (!collection) {
+      throw new NotFoundException('Collection not found')
+    }
+    if (collection.userId !== user.id && user.role !== UserRole.ADMIN) {
+      throw new ForbiddenException("You don't have permission to edit this resource")
+    }
+    return this.prisma.collection.update({
       where: { id },
       data: updateCollectionDto,
     })
+  }
+  async remove(id: string, user: User) {
+    const collection = await this.prisma.collection.findUnique({ where: { id } })
     if (!collection) {
       throw new NotFoundException('Collection not found')
     }
-    return collection
-  }
-  remove(id: string) {
-    try {
-      this.prisma.collection.delete({ where: { id } })
-    } catch (e) {
-      Logger.warn('Collection could not be deleted with id: ' + id)
+    if (collection.userId !== user.id && user.role !== UserRole.ADMIN) {
+      throw new ForbiddenException("You don't have permission to delete this resource")
     }
+    return this.prisma.collection.delete({ where: { id } })
   }
 }
